Time: 1 ms (100%), Space: 44.3 MB (10.65%) - LeetHub

+++ 3498-reverse-degree-of-a-string/3498-reverse-degree-of-a-string.java
@@ -1,10 +1,9 @@
 class Solution {
     public int reverseDegree(String s) {
-     int tot=0;
-     for(int i=0;i<s.length();i++){
-        tot+=(int)('z'-s.charAt(i)+1)*(i+1);
-       // System.out.println(tot);
-     }
-     return tot;
+        int tot=0;
+        for(int i=0;i<s.length();i++){
+            tot+=('z'-s.charAt(i)+1)*(i+1);
+        }
+        return tot;
     }
 }
